# openapi.yaml
openapi: 3.0.3
info:
  title: VK-Service API
  description: |
    File storage service with multi-provider support (Supabase S3, Google Drive).
    Provides user management, file uploads/downloads, and storage quota tracking.
  version: 1.0.0
  contact:
    name: VK-Service
    url: https://github.com/yourusername/vk-service
servers:
  - url: https://your-service.run.app/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development server
...
components:
  securitySchemes:
    ServiceSecret:
      type: apiKey
      in: header
      name: X-KV-SECRET
      description: Service secret key for protected endpoints

    UploadToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for file uploads (use X-Upload-Token header)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: string
          description: Additional error details
      required:
        - error
...
    FileMetadata:
      type: object
      properties:
        file_id:
          type: string
          example: 1a2b3c4d5e6f7890
          description: Unique file identifier (format varies by provider)
        file_name:
          type: string
          example: document.pdf
        mime_type:
          type: string
          example: application/pdf
        size:
          type: integer
          format: int64
          minimum: 0
          example: 1048576
          description: File size in bytes
        user_id:
          type: string
          format: uuid
          nullable: true
          example: 550e8400-e29b-41d4-a716-446655440000
          description: User ID for permanent files, null for temporal files
        description:
          type: string
          nullable: true
          example: Important document
        server_id:
          type: string
          format: uuid
          example: ff664d2f-cbc6-4436-81bf-e7ddff81f19d
        uploaded_at:
          type: string
          format: date-time
          example: 2025-12-15T16:00:00Z
        download_count:
          type: integer
          minimum: 0
          example: 3
        last_access:
          type: string
          format: date-time
          example: 2025-12-15T16:30:00Z
        delete_at:
          type: string
          format: date-time
          nullable: true
          example: 2025-12-15T17:00:00Z
          description: Deletion time for temporal files, null for permanent files
      required:
        - file_id
        - file_name
        - mime_type
        - size
...
    UpdateFileRequest:
      type: object
      properties:
        filename:
          type: string
          example: renamed-document.pdf
          description: New filename for the file
...
  parameters:
...
    FileIdPath:
      name: file_id
      in: path
      required: true
      schema:
        type: string
      description: File unique identifier
      example: 1a2b3c4d5e6f7890
  responses:
    BadRequest:
      description: Bad request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid request format
            code: BAD_REQUEST
...
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
            code: NOT_FOUND
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
            code: INTERNAL_ERROR
paths:
...
  /files/{file_id}:
    get:
      tags:
        - Files
      summary: Get file metadata
      description: Retrieve file metadata without downloading content
      operationId: getFileMetadata
      parameters:
        - $ref: '#/components/parameters/FileIdPath'
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadata'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Files
      summary: Update file metadata
      description: Update file metadata (e.g., rename file)
      operationId: updateFileMetadata
      parameters:
        - $ref: '#/components/parameters/FileIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFileRequest'
      responses:
        '200':
          description: File metadata updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadata'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Files
      summary: Delete file
      description: Delete a file from storage permanently
      operationId: deleteFile
      parameters:
        - $ref: '#/components/parameters/FileIdPath'
      responses:
        '204':
          description: File deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...